Move Testgrid Protobufs to dedicated folder
