+++ experiment/resultstore/BUILD.bazel
@@ -25,7 +25,7 @@ go_library(
     deps = [
         "//prow/flagutil:go_default_library",
         "//prow/logrusutil:go_default_library",
-        "//testgrid/config:go_default_library",
+        "//testgrid/data/config:go_default_library",
         "//testgrid/metadata:go_default_library",
         "//testgrid/resultstore:go_default_library",
         "//testgrid/util/gcs:go_default_library",
