+++ experiment/resultstore/main.go
@@ -35,7 +35,7 @@ import (
 
 	"k8s.io/test-infra/prow/flagutil"
 	"k8s.io/test-infra/prow/logrusutil"
-	"k8s.io/test-infra/testgrid/config"
+	"k8s.io/test-infra/testgrid/data/config"
 	"k8s.io/test-infra/testgrid/metadata"
 	"k8s.io/test-infra/testgrid/resultstore"
 	"k8s.io/test-infra/testgrid/util/gcs"
