+++ prow/spyglass/BUILD.bazel
@@ -19,7 +19,7 @@ go_test(
         "//prow/kube:go_default_library",
         "//prow/pod-utils/downwardapi:go_default_library",
         "//prow/spyglass/lenses:go_default_library",
-        "//testgrid/config:go_default_library",
+        "//testgrid/data/config:go_default_library",
         "//vendor/cloud.google.com/go/storage:go_default_library",
         "//vendor/github.com/fsouza/fake-gcs-server/fakestorage:go_default_library",
         "//vendor/github.com/sirupsen/logrus:go_default_library",
@@ -64,7 +64,7 @@ go_library(
         "//prow/deck/jobs:go_default_library",
         "//prow/pod-utils/gcs:go_default_library",
         "//prow/spyglass/lenses:go_default_library",
-        "//testgrid/config:go_default_library",
+        "//testgrid/data/config:go_default_library",
         "//testgrid/metadata:go_default_library",
         "//testgrid/util/gcs:go_default_library",
         "//vendor/cloud.google.com/go/storage:go_default_library",
