+++ prow/spyglass/testgrid_test.go
@@ -19,7 +19,7 @@ package spyglass
 import (
 	"testing"
 
-	tgconf "k8s.io/test-infra/testgrid/config"
+	tgconf "k8s.io/test-infra/testgrid/data/config"
 )
 
 func TestFindQuery(t *testing.T) {
