+++ testgrid/BUILD.bazel
@@ -14,13 +14,13 @@ filegroup(
         "//testgrid/cmd/entomologist:all-srcs",
         "//testgrid/cmd/summarizer:all-srcs",
         "//testgrid/cmd/updater:all-srcs",
-        "//testgrid/config:all-srcs",
+        "//testgrid/data/config:all-srcs",
+        "//testgrid/data/issue_state:all-srcs",
+        "//testgrid/data/state:all-srcs",
+        "//testgrid/data/summary:all-srcs",
         "//testgrid/images:all-srcs",
-        "//testgrid/issue_state:all-srcs",
         "//testgrid/metadata:all-srcs",
         "//testgrid/resultstore:all-srcs",
-        "//testgrid/state:all-srcs",
-        "//testgrid/summary:all-srcs",
         "//testgrid/util/gcs:all-srcs",
     ],
     tags = ["automanaged"],
