+++ testgrid/cmd/configurator/prow.go
@@ -18,7 +18,7 @@ package main
 
 import (
 	"fmt"
-	"k8s.io/test-infra/testgrid/config"
+	"k8s.io/test-infra/testgrid/data/config"
 	"path"
 	"strconv"
 	"strings"
