+++ testgrid/cmd/summarizer/BUILD.bazel
@@ -29,7 +29,7 @@ go_library(
     deps = [
         "//testgrid/cmd/summarizer/response:go_default_library",
         "//testgrid/cmd/summarizer/summary:go_default_library",
-        "//testgrid/summary:go_default_library",
+        "//testgrid/data/summary:go_default_library",
         "//vendor/github.com/sirupsen/logrus:go_default_library",
         "//vendor/google.golang.org/grpc:go_default_library",
     ],
@@ -41,8 +41,8 @@ go_test(
     embed = [":go_default_library"],
     deps = [
         "//testgrid/cmd/summarizer/response:go_default_library",
-        "//testgrid/config:go_default_library",
-        "//testgrid/summary:go_default_library",
+        "//testgrid/data/config:go_default_library",
+        "//testgrid/data/summary:go_default_library",
         "//vendor/github.com/google/go-cmp/cmp:go_default_library",
     ],
 )
