+++ testgrid/cmd/summarizer/example/BUILD.bazel
@@ -11,8 +11,8 @@ go_library(
     deps = [
         "//testgrid/cmd/summarizer/response:go_default_library",
         "//testgrid/cmd/summarizer/summary:go_default_library",
-        "//testgrid/config:go_default_library",
-        "//testgrid/summary:go_default_library",
+        "//testgrid/data/config:go_default_library",
+        "//testgrid/data/summary:go_default_library",
         "//vendor/github.com/sirupsen/logrus:go_default_library",
         "//vendor/google.golang.org/grpc:go_default_library",
         "//vendor/google.golang.org/grpc/credentials:go_default_library",
