+++ testgrid/cmd/summarizer/response/BUILD.bazel
@@ -6,8 +6,8 @@ go_library(
     importpath = "k8s.io/test-infra/testgrid/cmd/summarizer/response",
     visibility = ["//visibility:public"],
     deps = [
-        "//testgrid/config:go_default_library",
-        "//testgrid/summary:go_default_library",
+        "//testgrid/data/config:go_default_library",
+        "//testgrid/data/summary:go_default_library",
         "//vendor/github.com/golang/protobuf/proto:go_default_library",
     ],
 )
