+++ testgrid/cmd/summarizer/summary/BUILD.bazel
@@ -21,7 +21,7 @@ go_library(
     visibility = ["//visibility:public"],
     deps = [
         "//testgrid/cmd/summarizer/response:go_default_library",
-        "//testgrid/summary:go_default_library",
+        "//testgrid/data/summary:go_default_library",
         "//vendor/github.com/golang/protobuf/proto:go_default_library",
         "//vendor/google.golang.org/grpc:go_default_library",
     ],
