+++ testgrid/cmd/updater/BUILD.bazel
@@ -19,9 +19,9 @@ go_library(
     importpath = "k8s.io/test-infra/testgrid/cmd/updater",
     visibility = ["//visibility:private"],
     deps = [
-        "//testgrid/config:go_default_library",
+        "//testgrid/data/config:go_default_library",
+        "//testgrid/data/state:go_default_library",
         "//testgrid/metadata/junit:go_default_library",
-        "//testgrid/state:go_default_library",
         "//testgrid/util/gcs:go_default_library",
         "//vendor/cloud.google.com/go/storage:go_default_library",
         "//vendor/github.com/golang/protobuf/proto:go_default_library",
@@ -48,8 +48,8 @@ go_test(
     srcs = ["main_test.go"],
     embed = [":go_default_library"],
     deps = [
+        "//testgrid/data/state:go_default_library",
         "//testgrid/metadata/junit:go_default_library",
-        "//testgrid/state:go_default_library",
         "//vendor/github.com/golang/protobuf/proto:go_default_library",
     ],
 )
