+++ testgrid/cmd/updater/main_test.go
@@ -22,8 +22,8 @@ import (
 
 	"github.com/golang/protobuf/proto"
 
+	"k8s.io/test-infra/testgrid/data/state"
 	"k8s.io/test-infra/testgrid/metadata/junit"
-	"k8s.io/test-infra/testgrid/state"
 )
 
 func TestExtractRows(t *testing.T) {
