+++ testgrid/data/config/BUILD.bazel
@@ -6,7 +6,7 @@ go_library(
         "config.go",
         "config.pb.go",
     ],
-    importpath = "k8s.io/test-infra/testgrid/config",
+    importpath = "k8s.io/test-infra/testgrid/data/config",
     visibility = ["//visibility:public"],
     deps = [
         "//testgrid/util/gcs:go_default_library",
