+++ testgrid/data/summary/BUILD.bazel
@@ -16,8 +16,8 @@ filegroup(
 
 go_library(
     name = "go_default_library",
-    srcs = ["issue_state.pb.go"],
-    importpath = "k8s.io/test-infra/testgrid/issue_state",
+    srcs = ["summary.pb.go"],
+    importpath = "k8s.io/test-infra/testgrid/data/summary",
     visibility = ["//visibility:public"],
     deps = ["//vendor/github.com/golang/protobuf/proto:go_default_library"],
 )
